fix match comparator usage
If update_counter was true the match_key is the packet, and
the table match rule matches for that packet. On the other hand,
if update_counter was false, the match_key is match rule for
FLOW_MOD or MULTIPART FLOW stats. In that case, match_key
matches for the table match rule.

Signed-off-by: Hiroaki KAWAI <kawai@stratosphere.co.jp>